requirements.txt
absl-py==2.3.1
annotated-types==0.7.0
asttokens
astunparse==1.6.3
Bottleneck
Brotli
brotlicffi
cachetools==5.5.2
certifi
cffi
charset-normalizer
colorama
comm
contourpy
cycler
debugpy
decorator
dysts==0.96
exceptiongroup
executing
filelock
flatbuffers==25.2.10
fonttools
gast==0.4.0
gluonts==0.16.2
gmpy2
google-auth==2.40.3
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.73.1
h5py==3.14.0
idna
imbalanced-learn
importlib_metadata
ipykernel
ipython
jedi
Jinja2
joblib
jupyter_client
jupyter_core
keras==2.10.0
Keras-Preprocessing==1.1.2
keras-tuner==1.4.8
kiwisolver
libclang==18.1.1
llvmlite
Markdown==3.8.2
markdown-it-py==3.0.0
MarkupSafe
matplotlib==3.10.0
matplotlib-inline
mdurl==0.1.2
mkl-service==2.4.0
mkl_fft
mkl_random
ml_dtypes==0.5.1
mpmath
namex==0.1.0
nest_asyncio
networkx
numba
numexpr
numpy==1.26.4
oauthlib==3.3.1
opt_einsum==3.4.0
optree==0.16.0
packaging
pandas==2.2.3
parso
pickleshare
pillow
platformdirs
prompt_toolkit
protobuf==3.19.6
psutil
pure_eval
pyarrow==20.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser
pydantic==2.11.7
pydantic_core==2.33.2
Pygments
pyparsing
PyQt6==6.7.1
PyQt6_sip
pyreco==1.0.3
PySocks
python-dateutil
pytz
pywin32
PyYAML
pyzmq
requests
requests-oauthlib==2.0.0
rich==14.0.0
rsa==4.9.1
scikit-learn==1.6.1
scipy==1.15.3
seaborn
sip
six
sklearn-compat
stack_data
sympy
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.10.1
tensorflow-estimator==2.10.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==3.1.0
threadpoolctl
tomli
toolz==0.12.1
torch==2.5.1
torchaudio==2.5.1
torchvision==0.20.1
tornado
tqdm==4.67.1
traitlets
typing-inspection==0.4.1
typing_extensions
tzdata
unicodedata2
urllib3
wcwidth
Werkzeug==3.1.3
win-inet-pton
wrapt==1.17.2
zipp